Ajustar o code para receber numero decimal na posologia

+++ src/org/celllife/idart/gui/packaging/NewPatientPackaging.java
@@ -1618,7 +1618,8 @@ public class NewPatientPackaging extends GenericFormGui implements iDARTChangeLi
      */
     private boolean fieldsOkay(java.util.List<PackageDrugInfo> allPackagedDrugsList) {
         Patient patient = PatientManager.getPatient(getHSession(), localPatient.getId());
-        int amountperPackage = (int) (allPackagedDrugsList.get(0).getDispensedQty() / (allPackagedDrugsList.get(0).getTimesPerDay() * Integer.parseInt(allPackagedDrugsList.get(0).getAmountPerTime()))) / 7;
+        int amountperPackage = (int) (allPackagedDrugsList.get(0).getDispensedQty() / (allPackagedDrugsList.get(0).getTimesPerDay() * 
+        		Double.parseDouble(allPackagedDrugsList.get(0).getAmountPerTime()))) / 7;
 
         if (localPatient.getCurrentPrescription().getDuration() != newPack.getWeekssupply()) {
             MessageBox mb = new MessageBox(getShell(), SWT.ICON_QUESTION | SWT.YES | SWT.NO);
@@ -3937,5 +3938,4 @@ public class NewPatientPackaging extends GenericFormGui implements iDARTChangeLi
             }
         }
     }
-
 }
